Remove the redundant param to update a single plugin, change update to upgrade
